requirements.txt
nltk
streamlit
sentence_transformers
requests
pyjwt
typing
streamlit_option_menu
